ZainWireless/Backend/requirements.txt
aiodns==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
ccxt==4.4.33
certifi==2025.1.31
cffi==1.17.1
channels==4.0.0
charset-normalizer==3.4.1
cryptography==43.0.3
dj-rest-auth==5.0.2
Django==5.1.6
django-allauth==0.63.3
django-cors-headers==4.7.0
django-filter==25.1
django-nextjs==2.4.0
django-rest-authtoken==2.1.4
django-rest-passwordreset==1.4.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
gunicorn==23.0.0
idna==3.10
multidict==6.0.4
mysqlclient==2.2.4
numpy==2.1.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycares==4.4.0
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2023.3.post1
requests==2.32.3
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.5.3
stripe==11.5.0
tweepy==4.14.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websockets==12.0
yarl==1.9.4
